move the Memoizer class to the model.api package
to keep the api classes contained within a single package. This will make
externalizing into a bundle easier in the future.
